requirements.txt
pytest
pyyaml
scikit-learn
pandas
numpy